seoul-live/requirements.txt
requests
supabase
streamlit
python-dotenv
plotly
pandas
